Changed the PrintPreview flag to true.
For the rendering of voice parts the PrintPreview flag was set to false.
This caused all the notes of a voice to be squeezed into one line when
generating a PDF and with some of the variant versions in the parts window
view.

The single voice print issue is now solved for 2 cases:

1. When generating a PDF through the Menu "File > Generate
PDF.. > Parts > Generate..."
2. When opening the parts window through the Menu "View > View parts
window".